Broke HS.BuildStreamNodes() into 3 separate methods to clean up the implementation. Reach now contains one initial boundary condition node at headwater and optionally uneven node at mouth.
Remove BoundCond class definition. Boundary conditions are now
only in T_bc, Q_bc and C_bc. These attributes are set to None
in all but the boundary StreamNode.

Created StreamNode.GetAttributes(zone=False) and 
GetZoneAttributes(), which will return a dictionary of the
attributes of the StreamNode instance (optionally with zone),
and the StreamNode.Zone instance.

git-svn-id: file://deqhq1/tmdl/Subversion/HeatSource/HeatSource@45 ffc44b20-e41f-394f-8689-23e218f545e0