remove radial sampling option and various print statements, add setup.py (#1)
* Remove radial sampling option and various print statements,

This set of changes removes testing features that were included in the main branch that allows a user defined number of radial samples. This feature was removed and returns to the standard 7 direction sample method. For the most part the code was returned to the same state as commit 70406ef with the exception of the following:

Various print statements were removed.  These statements significantly increased model run times.

The writing of errors to text files in the root C directory was also removed. The C directory is read only in newer versions windows. An error is caused by trying to write there and the original error message was lost.

The heat source model version is now printed on every output text file for record keeping.

* Delete run_hs.py

* Fixed the print message for unstable timesteps.

* Update README.md

* switch back to version 8.0.8

* Create setup.py

* Revert "Update README.md"

This reverts commit 92738f6689151d117b8345e3e483d6b73b11c7b8.

* clairifed setup folder structure